Refactor MCP rule source helpers
